requirements.txt
Django>=4.0.5
djangorestframework>=3.13.1
djangorestframework-simplejwt>=5.2.0
psycopg2-binary>=2.9.3
gunicorn>=20.1.0